Add files via upload
A Library Assistant is a self-billing system that can help in many book devices and shops or any other self-billing policy.
 It is basically a system that deals with the current high-speed information processing movement, by dealing with each client easily and only once for each duration. Besides, keep tracing the total process and delete deponed in the user choice. 
once the Bill printed all the data disappear except the total tracing. The idea of self-billing will reduce the effort and time. 
In order to keep tracing the continuous process of borrowing from the machine, I used the stack, the functionality of first in last out ease saving the semi-old data, which is not important for the user but pretty valuable for the system owner.
It's also useful in the privacy situation if the user doesn’t want to keep his data, once the bill printed everything will go. 
On the other hand, due to its fast insertion and deletion, it will be better than the usual array process, with less complexity since each process will take only O(1)!
The ArrayList will provide an efficient solution to the space problem in the bigger projects …